requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.7.0
python-dotenv>=1.0.0
logfire>=3.14.0
httpx>=0.24.0
openai>=1.0.0
